Change weight log attribute weight to integer

+++ db/migrate/20190721203737_create_weight_logs.rb
@@ -1,7 +1,7 @@
 class CreateWeightLogs < ActiveRecord::Migration[5.2]
   def change
     create_table :weight_logs do |t|
-      t.decimal :weight
+      t.integer :weight
       t.references :client, foreign_key: true
 
       t.timestamps
